desktop: Move debug backend to desktop
