+++ desktop/Cargo.toml
@@ -27,6 +27,7 @@ rfd = "0.10.0"
 anyhow = "1.0"
 bytemuck = "1.13.0"
 os_info = { version = "3", default-features = false }
+tungstenite = { version = "0.18.0"}
 
 # Deliberately held back to match tracy client used by profiling crate
 tracing-tracy = { version = "=0.10.0", optional = true }
