+++ desktop/src/main.rs
@@ -13,6 +13,7 @@ mod navigator;
 mod storage;
 mod task;
 mod ui;
+mod debug;
 
 use crate::custom_event::RuffleEvent;
 use crate::executor::GlutinAsyncExecutor;
@@ -42,6 +43,7 @@ use winit::event::{
 };
 use winit::event_loop::{ControlFlow, EventLoop, EventLoopBuilder};
 use winit::window::{Fullscreen, Icon, Window, WindowBuilder};
+use crate::debug::WebsocketDebugBackend;
 
 thread_local! {
     static CALLSTACK: RefCell<Option<StaticCallstack>> = RefCell::default();
@@ -297,7 +299,7 @@ impl App {
             .with_load_behavior(opt.load_behavior)
             .with_spoofed_url(opt.spoof_url.clone().map(|url| url.to_string()))
             .with_player_version(opt.player_version)
-            .with_debugger(ruffle_core::backend::debug::WebsocketDebugBackend::new());
+            .with_debugger(WebsocketDebugBackend::new());
 
         let player = builder.build();
